requirements.txt
python-telegram-bot[job-queue]==21.0
gspread==6.0.0
google-auth==2.27.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
